Fix checking of base_uri after evaluation, move test cases to functaional tests
