+++ Tests/Unit/Components/WebService/Rest/RestConfigurableProducerTest.php
@@ -15,7 +15,6 @@ use JMS\Serializer\SerializerInterface;
 use Smartbox\CoreBundle\Tests\Fixtures\Entity\SerializableThing;
 use Smartbox\Integration\FrameworkBundle\Components\WebService\ConfigurableWebserviceProtocol;
 use Smartbox\Integration\FrameworkBundle\Components\WebService\Rest\Exceptions\RecoverableRestException;
-use Smartbox\Integration\FrameworkBundle\Components\WebService\Rest\Exceptions\RestException;
 use Smartbox\Integration\FrameworkBundle\Components\WebService\Rest\Exceptions\UnrecoverableRestException;
 use Smartbox\Integration\FrameworkBundle\Components\WebService\Rest\RestConfigurableProducer;
 use Smartbox\Integration\FrameworkBundle\Components\WebService\Rest\RestConfigurableProtocol;
@@ -131,99 +130,6 @@ class RestConfigurableProducerTest extends \PHPUnit\Framework\TestCase
         /** @var RequestException $originalException */
         $originalException = $exception->getPrevious();
         $this->assertEquals($statusCode, $originalException->getResponse()->getStatusCode());
-    }
-
-    public function getCasesForBaseUri()
-    {
-        return [
-            'default' => [
-                'base_uri' => 'http://someservice.com/api/',
-                'uri' => 'v0/put/cats',
-                'overRideBaseUri' => false,
-                'expectedHost' => 'someservice.com',
-                'expectedPath' => '/api/v0/put/cats',
-            ],
-            'override' => [
-                'base_uri' => 'http://someservice.com/api/',
-                'uri' => 'v0/put/cats',
-                'overRideBaseUri' => 'http://dev.machine/api/',
-                'expectedHost' => 'dev.machine',
-                'expectedPath' => '/api/v0/put/cats',
-            ],
-            'override but ignore null' => [
-                'base_uri' => 'http://someservice.com/api/',
-                'uri' => 'v0/put/cats',
-                'overRideBaseUri' => null,
-                'expectedHost' => 'someservice.com',
-                'expectedPath' => '/api/v0/put/cats',
-            ],
-        ];
-    }
-
-    /**
-     * @param $baseUri
-     * @param $requestUri
-     * @param $overRideBaseUri
-     * @param $expectedHost
-     * @param $expectedPath
-     * @dataProvider getCasesForBaseUri
-     */
-    public function testCanOverrideBaseUrI($baseUri, $requestUri, $overRideBaseUri, $expectedHost, $expectedPath)
-    {
-        $mockHandler = new MockHandler([
-            new Response(200, ['X-Foo' => 'Bar']),
-        ]);
-
-        $container = [];
-        $history = Middleware::history($container);
-        $stack = HandlerStack::create($mockHandler);
-        $stack->push($history);
-
-        $client = new Client(['handler' => $stack]);
-
-        $producer = new RestConfigurableProducer();
-        $producer->setHttpClient($client);
-        $producer->setEvaluator($this->evaluator);
-        $producer->setSerializer($this->serializer);
-        $producer->setConfHelper($this->helper);
-        $producer->setConfHelper($this->helper);
-        $producer->setEventDispatcher($this->eventDispatcher);
-
-        $actionParams = [
-            RestConfigurableProducer::REQUEST_NAME => 'someRequest',
-            RestConfigurableProducer::REQUEST_HTTP_VERB => 'POST',
-            RestConfigurableProducer::REQUEST_BODY => ['hello' => 'world'],
-            RestConfigurableProducer::REQUEST_URI => $requestUri,
-            RestConfigurableProducer::DISPLAY_RESPONSE_ERROR => false,
-        ];
-
-        $options = [
-            RestConfigurableProtocol::OPTION_BASE_URI => $baseUri,
-            RestConfigurableProtocol::OPTION_ENCODING => RestConfigurableProtocol::ENCODING_JSON,
-            ConfigurableWebserviceProtocol::OPTION_CONNECT_TIMEOUT => 10,
-            ConfigurableWebserviceProtocol::OPTION_TIMEOUT => 10,
-            RestConfigurableProtocol::OPTION_HEADERS => [],
-            RestConfigurableProtocol::OPTION_AUTH => false,
-        ];
-        $context = [
-            'vars' => [],
-            'msg'  => new Message(new SerializableThing(), [], new Context()),
-            'exchange' => new Exchange(),
-        ];
-
-        //Decide to override the base_uri or not
-        if (false !== $overRideBaseUri) {
-            $actionParams['base_uri'] = $overRideBaseUri;
-        }
-
-        $response = $producer->executeStep(RestConfigurableProducer::STEP_REQUEST, $actionParams, $options, $context);
-        $this->assertTrue($response, 'The producer should return true to say it has completed the Request Step');
-
-        /** @var Request $request */
-        $request = $container[0]['request'];
-
-        $this->assertEquals($expectedHost, $request->getUri()->getHost());
-        $this->assertEquals($expectedPath, $request->getUri()->getPath());
     }
 
     public function getExpectedExceptionsForStatusCodes()
